pyproject.toml
[project]
name = "news-digest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
